backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pytest
flake8
black
isort
openai
fastapi
uvicorn
pydantic
requests
python-dotenv
pytest
flake8
black
isort
openai
anthropic
google-generativeai
voyageai